Update Air_Quality.py
change in line 33: int(aqi) to float(aqi)
the data format of the website was changed.

+++ Assignment_1_9.25/Air_Quality.py
@@ -30,7 +30,7 @@ for citydata in airdata['results'][:20]:
     citylabel = f"{cityname}, {path}" if path else cityname
     
     cities.append(citylabel)
-    aqis.append(int(aqi))  
+    aqis.append(float(aqi))  
 
 
 # Change the Color
